LPS-124006 Extract ImageSelectorDescription
